requirements.txt
langchain
langchain-openai
langchain-community
youtube-transcript-api
faiss-cpu
tiktoken
python-dotenv
